Rename components
* access -> publisher
* manage -> manager
* migrate -> migrator
