+++ .goreleaser.yaml
@@ -1,9 +1,9 @@
 ---
 version: 2
 builds:
-  - id: archived-access
-    main: ./cmd/access
-    binary: archived-access
+  - id: archived-publisher
+    main: ./cmd/publisher
+    binary: archived-publisher
     ldflags:
       - -s -w -X main.appVersion={{.Version}} -X main.buildTimestamp={{.Date}}
     env:
@@ -50,9 +50,9 @@ builds:
     goarm: ["6", "7"]
     gomips: ["hardfloat", "softfloat"]
     mod_timestamp: "{{ .CommitTimestamp }}"
-  - id: archived-manage
-    main: ./cmd/manage
-    binary: archived-manage
+  - id: archived-manager
+    main: ./cmd/manager
+    binary: archived-manager
     ldflags:
       - -s -w -X main.appVersion={{.Version}} -X main.buildTimestamp={{.Date}}
     env:
@@ -65,9 +65,9 @@ builds:
     goamd64: ["v1", "v2", "v3"]
     goarm: ["7"]
     mod_timestamp: "{{ .CommitTimestamp }}"
-  - id: archived-migrate
-    main: ./cmd/migrate
-    binary: archived-migrate
+  - id: archived-migrator
+    main: ./cmd/migrator
+    binary: archived-migrator
     ldflags:
       - -s -w -X main.appVersion={{.Version}} -X main.buildTimestamp={{.Date}}
     env:
