+++ Dockerfile.migrator
@@ -6,7 +6,7 @@ RUN apk add --update --no-cache \
 FROM scratch
 
 COPY --from=certificates /etc/ssl/cert.pem /etc/ssl/cert.pem
-COPY --chmod=0755 --chown=root:root dist/archived-migrate_linux_amd64_v3/archived-migrate /archived-migrate
+COPY --chmod=0755 --chown=root:root dist/archived-migrator_linux_amd64_v3/archived-migrator /archived-migrator
 COPY --chmod=0644 --chown=root:root repositories/metadata/postgresql/migrations/sql /migrations/sql
 
-ENTRYPOINT [ "/archived-migrate" ]
+ENTRYPOINT [ "/archived-migrator" ]
