+++ Dockerfile.publisher
@@ -1,0 +1,14 @@
+FROM alpine:3.20.0 AS certificates
+
+RUN apk add --update --no-cache \
+  ca-certificates=20240705-r0
+
+FROM scratch
+
+COPY --from=certificates /etc/ssl/cert.pem /etc/ssl/cert.pem
+COPY --chmod=0755 --chown=root:root dist/archived-publisher_linux_amd64_v3/archived-publisher /archived-publisher
+COPY --chmod=0644 --chown=root:root presenter/publisher/html/templates /templates
+
+ENV HTML_TEMPLATE_DIR=/templates
+
+ENTRYPOINT [ "/archived-publisher" ]
