+++ presenter/publisher/html/templates/object-list.html
