+++ presenter/publisher/html/testdata/objects.html.sample
